memberof: free delete operation payload once done
memberof: fix calculation of replaced members

We were skipping the check on the next value in the added list when a match
was found for the currentr value being checked.

memberof: free delete operation apyload once done

Large memberof delete operations can cause quite a number of searches
and the results are attached to a delop operation structure.
Make sure we free this payload once the operation is done and these
results are not used anymore so that we get a smaller total memory footprint.