Special-case LDAP_SIZELIMIT_EXCEEDED
Previous version of the SSSD did not abort the async LDAP search
operation on errors. In cases where the request ended in progress, such
as when the paging was very strictly limited, the old versions at least
returned partial data.

This patch special-cases the LDAP_SIZELIMIT_EXCEEDED error to avoid a
user-visible regression.

https://fedorahosted.org/sssd/ticket/1322