Release mutex if a thread gets cancelled
Our code in the client calls several thread cancellation points after a
mutex is locked. If the thread gets cancelled after the mutex has been
acquired, it is never released, causing a deadlock.

This patch adds a cleanup routine registered with pthread_cleanup_push,
which would be invoked on thread cancelation. If the thread is canceled,
the cleanup routine releases the mutex and closes the connection to the
responder.

https://fedorahosted.org/sssd/ticket/1460